pyproject.toml
[project]
name = "romav2"
version = "2.0.0"
description = "A very good dense feature matcher."
readme = "README.md"
authors = [
    { name = "Johan Edstedt", email = "johan.edstedt@liu.se" }
]
requires-python = ">=3.10"
dependencies = [
    "dataclasses>=0.8",
    "einops>=0.8.1",
    "pillow>=12.0.0",
    "rich>=14.2.0",
    "torch",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[project.optional-dependencies]
fused-local-corr = [
    "fused-local-corr>=0.2.4 ; sys_platform == 'linux'",
]
eval = [
    "kornia>=0.8.2",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "wandb>=0.23.0",
    "wxbs-benchmark>=0.0.4",
]
dev = [
    "slurm-util>=0.2.7",
    "ruff>=0.14.5",
]
